lambda-functions/docker-lambda/requirements.txt
requests==2.28.2
beautifulsoup4==4.12.2
pillow==9.5.0
pydantic==2.6.1